feat: refactor chat msg fetching and order contacts by most recent msg
