+++ packages/legacy/core/App/screens/Chat.tsx
@@ -1,18 +1,9 @@
-import {
-  BasicMessageRecord,
-  BasicMessageRepository,
-  CredentialExchangeRecord,
-  CredentialState,
-  ProofExchangeRecord,
-  ProofState,
-} from '@aries-framework/core'
+import { BasicMessageRepository } from '@aries-framework/core'
 import { useAgent, useBasicMessagesByConnectionId, useConnectionById } from '@aries-framework/react-hooks'
-import { isPresentationReceived } from '@hyperledger/aries-bifold-verifier'
 import { useIsFocused, useNavigation } from '@react-navigation/core'
 import { StackNavigationProp, StackScreenProps } from '@react-navigation/stack'
-import React, { Fragment, useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { Linking, Text } from 'react-native'
 import { GiftedChat, IMessage } from 'react-native-gifted-chat'
 import { SafeAreaView } from 'react-native-safe-area-context'
 
@@ -20,24 +11,15 @@ import InfoIcon from '../components/buttons/InfoIcon'
 import { renderComposer, renderInputToolbar, renderSend } from '../components/chat'
 import ActionSlider from '../components/chat/ActionSlider'
 import { renderActions } from '../components/chat/ChatActions'
-import { ChatEvent } from '../components/chat/ChatEvent'
-import { ChatMessage, ExtendedChatMessage, CallbackType } from '../components/chat/ChatMessage'
+import { ChatMessage } from '../components/chat/ChatMessage'
 import { useNetwork } from '../contexts/network'
 import { useStore } from '../contexts/store'
 import { useTheme } from '../contexts/theme'
-import { useCredentialsByConnectionId } from '../hooks/credentials'
-import { useProofsByConnectionId } from '../hooks/proofs'
+import { useChatMessagesByConnection } from '../hooks/chat-messages'
 import { Role } from '../types/chat'
 import { BasicMessageMetadata, basicMessageCustomMetadata } from '../types/metadata'
 import { RootStackParams, ContactStackParams, Screens, Stacks } from '../types/navigators'
-import {
-  getConnectionName,
-  getCredentialEventLabel,
-  getCredentialEventRole,
-  getMessageEventRole,
-  getProofEventLabel,
-  getProofEventRole,
-} from '../utils/helpers'
+import { getConnectionName } from '../utils/helpers'
 
 type ChatProps = StackScreenProps<ContactStackParams, Screens.Chat> | StackScreenProps<RootStackParams, Screens.Chat>
 
@@ -53,14 +35,11 @@ const Chat: React.FC<ChatProps> = ({ route }) => {
   const navigation = useNavigation<StackNavigationProp<RootStackParams | ContactStackParams>>()
   const connection = useConnectionById(connectionId)
   const basicMessages = useBasicMessagesByConnectionId(connectionId)
-  const credentials = useCredentialsByConnectionId(connectionId)
-  const proofs = useProofsByConnectionId(connectionId)
+  const chatMessages = useChatMessagesByConnection(connection)
   const isFocused = useIsFocused()
   const { assertConnectedNetwork, silentAssertConnectedNetwork } = useNetwork()
-  const [messages, setMessages] = useState<Array<ExtendedChatMessage>>([])
   const [showActionSlider, setShowActionSlider] = useState(false)
   const { ChatTheme: theme, Assets } = useTheme()
-  const { ColorPallet } = useTheme()
   const [theirLabel, setTheirLabel] = useState(getConnectionName(connection, store.preferences.alternateContactNames))
 
   // This useEffect is for properly rendering changes to the alt contact name, useMemo did not pick them up
@@ -90,183 +69,6 @@ const Chat: React.FC<ChatProps> = ({ route }) => {
       }
     })
   }, [basicMessages])
-
-  useEffect(() => {
-    const transformedMessages: Array<ExtendedChatMessage> = basicMessages.map((record: BasicMessageRecord) => {
-      const role = getMessageEventRole(record)
-      // eslint-disable-next-line
-      const linkRegex = /(?:https?\:\/\/\w+(?:\.\w+)+\S*)|(?:[\w\d\.\_\-]+@\w+(?:\.\w+)+)/gm
-      // eslint-disable-next-line
-      const mailRegex = /^[\w\d\.\_\-]+@\w+(?:\.\w+)+$/gm
-      const links = record.content.match(linkRegex) ?? []
-      const handleLinkPress = (link: string) => {
-        if (link.match(mailRegex)) {
-          link = 'mailto:' + link
-        }
-        Linking.openURL(link)
-      }
-      const msgText = (
-        <Text style={role === Role.me ? theme.rightText : theme.leftText}>
-          {record.content.split(linkRegex).map((split, i) => {
-            if (i < links.length) {
-              const link = links[i]
-              return (
-                <Fragment key={`${record.id}-${i}`}>
-                  <Text>{split}</Text>
-                  <Text
-                    onPress={() => handleLinkPress(link)}
-                    style={{ color: ColorPallet.brand.link, textDecorationLine: 'underline' }}
-                    accessibilityRole={'link'}
-                  >
-                    {link}
-                  </Text>
-                </Fragment>
-              )
-            }
-            return <Text key={`${record.id}-${i}`}>{split}</Text>
-          })}
-        </Text>
-      )
-      return {
-        _id: record.id,
-        text: record.content,
-        renderEvent: () => msgText,
-        createdAt: record.updatedAt || record.createdAt,
-        type: record.type,
-        user: { _id: role },
-      }
-    })
-
-    const callbackTypeForMessage = (record: CredentialExchangeRecord | ProofExchangeRecord) => {
-      if (
-        record instanceof CredentialExchangeRecord &&
-        (record.state === CredentialState.Done || record.state === CredentialState.OfferReceived)
-      ) {
-        return CallbackType.CredentialOffer
-      }
-
-      if (
-        (record instanceof ProofExchangeRecord && isPresentationReceived(record) && record.isVerified !== undefined) ||
-        record.state === ProofState.RequestReceived ||
-        (record.state === ProofState.Done && record.isVerified === undefined)
-      ) {
-        return CallbackType.ProofRequest
-      }
-
-      if (
-        record instanceof ProofExchangeRecord &&
-        (record.state === ProofState.PresentationSent || record.state === ProofState.Done)
-      ) {
-        return CallbackType.PresentationSent
-      }
-    }
-
-    transformedMessages.push(
-      ...credentials.map((record: CredentialExchangeRecord) => {
-        const role = getCredentialEventRole(record)
-        const userLabel = role === Role.me ? t('Chat.UserYou') : theirLabel
-        const actionLabel = t(getCredentialEventLabel(record) as any)
-
-        return {
-          _id: record.id,
-          text: actionLabel,
-          renderEvent: () => <ChatEvent role={role} userLabel={userLabel} actionLabel={actionLabel} />,
-          createdAt: record.updatedAt || record.createdAt,
-          type: record.type,
-          user: { _id: role },
-          messageOpensCallbackType: callbackTypeForMessage(record),
-          onDetails: () => {
-            const navMap: { [key in CredentialState]?: () => void } = {
-              [CredentialState.Done]: () => {
-                navigation.navigate(Stacks.ContactStack as any, {
-                  screen: Screens.CredentialDetails,
-                  params: { credential: record },
-                })
-              },
-              [CredentialState.OfferReceived]: () => {
-                navigation.navigate(Stacks.ContactStack as any, {
-                  screen: Screens.CredentialOffer,
-                  params: { credentialId: record.id },
-                })
-              },
-            }
-            const nav = navMap[record.state]
-            if (nav) {
-              nav()
-            }
-          },
-        }
-      })
-    )
-
-    transformedMessages.push(
-      ...proofs.map((record: ProofExchangeRecord) => {
-        const role = getProofEventRole(record)
-        const userLabel = role === Role.me ? t('Chat.UserYou') : theirLabel
-        const actionLabel = t(getProofEventLabel(record) as any)
-
-        return {
-          _id: record.id,
-          text: actionLabel,
-          renderEvent: () => <ChatEvent role={role} userLabel={userLabel} actionLabel={actionLabel} />,
-          createdAt: record.updatedAt || record.createdAt,
-          type: record.type,
-          user: { _id: role },
-          messageOpensCallbackType: callbackTypeForMessage(record),
-          onDetails: () => {
-            const toProofDetails = () => {
-              navigation.navigate(Stacks.ContactStack as any, {
-                screen: Screens.ProofDetails,
-                params: {
-                  recordId: record.id,
-                  isHistory: true,
-                  senderReview:
-                    record.state === ProofState.PresentationSent ||
-                    (record.state === ProofState.Done && record.isVerified === undefined),
-                },
-              })
-            }
-            const navMap: { [key in ProofState]?: () => void } = {
-              [ProofState.Done]: toProofDetails,
-              [ProofState.PresentationSent]: toProofDetails,
-              [ProofState.PresentationReceived]: toProofDetails,
-              [ProofState.RequestReceived]: () => {
-                navigation.navigate(Stacks.ContactStack as any, {
-                  screen: Screens.ProofRequest,
-                  params: { proofId: record.id },
-                })
-              },
-            }
-            const nav = navMap[record.state]
-            if (nav) {
-              nav()
-            }
-          },
-        }
-      })
-    )
-
-    const connectedMessage = connection
-      ? {
-          _id: 'connected',
-          text: `${t('Chat.YouConnected')} ${theirLabel}`,
-          renderEvent: () => (
-            <Text style={theme.rightText}>
-              {t('Chat.YouConnected')}
-              <Text style={[theme.rightText, theme.rightTextHighlighted]}> {theirLabel}</Text>
-            </Text>
-          ),
-          createdAt: connection.createdAt,
-          user: { _id: Role.me },
-        }
-      : undefined
-
-    setMessages(
-      connectedMessage
-        ? [...transformedMessages.sort((a: any, b: any) => b.createdAt - a.createdAt), connectedMessage]
-        : transformedMessages.sort((a: any, b: any) => b.createdAt - a.createdAt)
-    )
-  }, [basicMessages, credentials, proofs, theirLabel])
 
   const onSend = useCallback(
     async (messages: IMessage[]) => {
@@ -304,7 +106,7 @@ const Chat: React.FC<ChatProps> = ({ route }) => {
   return (
     <SafeAreaView edges={['bottom', 'left', 'right']} style={{ flex: 1, paddingTop: 20 }}>
       <GiftedChat
-        messages={messages}
+        messages={chatMessages}
         showAvatarForEveryMessage={true}
         alignTop
         renderAvatar={() => null}
